mod25/Projeto2/projeto/dao/VendaDAO.java
package mod25.Projeto2.projeto.dao;

import mod25.Projeto2.projeto.dao.generic.GenericDAO;
import mod25.Projeto2.projeto.domain.Venda;
import mod25.Projeto2.projeto.exceptions.TipoChaveNaoEncontradaException;

public class VendaDAO extends GenericDAO<Venda, String> implements IVendaDAO {

	@Override
	public Class<Venda> getTipoClasse() {
		return Venda.class;
	}

	@Override
	public void atualizarDados(Venda entity, Venda entityCadastrado) {
		entityCadastrado.setCodigo(entity.getCodigo());
		entityCadastrado.setStatus(entity.getStatus());
	}

	@Override
	public void excluir(String valor) {
		throw new UnsupportedOperationException("Operação não permitida");
	}

	@Override
	public void finalizarVenda(Venda venda) throws TipoChaveNaoEncontradaException {
		venda.setStatus(Venda.Status.CONCLUIDA);
		super.alterar(venda);
	}
	
	

}
